Add files via upload
This code defines a function `get_client_info` that takes an IP address as its argument and returns a dictionary of information about that client. It uses the `socket` library to create a TCP socket and connect to the client's IP address. Then, it sends a simple HTTP request to that IP address and gets a response. Finally, it parses the response to extract the HTTP status code and server type, as well as the client's hostname using the `gethostbyaddr` function.